requirements.txt
asgiref==3.4.1
boto3==1.18.16
botocore==1.21.16
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.2.6
django-heroku==0.3.1
django-import-export==2.5.0
django-storages==1.11.1
djangorestframework==3.12.4
et-xmlfile==1.1.0
gunicorn==20.1.0
jmespath==0.10.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.7
psycopg2==2.9.1
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.1
tablib==3.0.0
urllib3==1.26.6
whitenoise==5.3.0
xlrd==2.0.1
xlwt==1.3.0
